creo haber arreglado el lio
